+++ packages/react-learn-card/src/components/CertificateDisplayCard/CertificateDisplayCard.tsx
@@ -39,6 +39,7 @@ type CertificateDisplayCardProps = {
     showDetailsBtn?: boolean;
     formattedDisplayType?: string;
     customLinkedCredentialsComponent?: React.ReactNode;
+    customBodyContentSlot?: React.ReactNode;
 };
 
 export const CertificateDisplayCard: React.FC<CertificateDisplayCardProps> = ({
@@ -68,6 +69,7 @@ export const CertificateDisplayCard: React.FC<CertificateDisplayCardProps> = ({
     showDetailsBtn = false,
     formattedDisplayType,
     customLinkedCredentialsComponent,
+    customBodyContentSlot,
 }) => {
     const [_isFront, _setIsFront] = useState(isFrontOverride ?? true);
 
@@ -92,6 +94,7 @@ export const CertificateDisplayCard: React.FC<CertificateDisplayCardProps> = ({
                         isFront={isFront}
                         showDetailsBtn={showDetailsBtn}
                         formattedDisplayType={formattedDisplayType}
+                        customBodyContentSlot={customBodyContentSlot}
                     />
                 )}