+++ packages/learn-card-cli/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @learncard/cli
 
+## 3.3.146
+
+### Patch Changes
+
+-   Updated dependencies [[`a8ba030d48e75094fd64cd3da0725c3c0f468cf2`](https://github.com/learningeconomy/LearnCard/commit/a8ba030d48e75094fd64cd3da0725c3c0f468cf2)]:
+    -   @learncard/types@5.9.1
+    -   @learncard/init@2.1.13
+    -   @learncard/core@9.3.43
+    -   @learncard/didkit-plugin@1.5.36
+    -   @learncard/learn-cloud-plugin@2.2.9
+    -   @learncard/linked-claims-plugin@0.1.3
+    -   @learncard/open-badge-v2-plugin@1.0.11
+    -   @learncard/simple-signing-plugin@1.0.36
+
 ## 3.3.145
 
 ### Patch Changes