+++ packages/plugins/crypto/CHANGELOG.md
@@ -1,5 +1,12 @@
 # learn-card-core
 
+## 1.0.54
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/core@9.3.43
+
 ## 1.0.53
 
 ### Patch Changes