+++ packages/plugins/vc/CHANGELOG.md
@@ -1,5 +1,14 @@
 # learn-card-core
 
+## 1.2.6
+
+### Patch Changes
+
+-   Updated dependencies [[`a8ba030d48e75094fd64cd3da0725c3c0f468cf2`](https://github.com/learningeconomy/LearnCard/commit/a8ba030d48e75094fd64cd3da0725c3c0f468cf2)]:
+    -   @learncard/types@5.9.1
+    -   @learncard/core@9.3.43
+    -   @learncard/didkit-plugin@1.5.36
+
 ## 1.2.5
 
 ### Patch Changes